requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
torchaudio==0.8.1
torchvision==0.9.1+cpu
alabaster==0.7.12
attrs==21.2.0
backcall==0.2.0
click==8.0.1
colorama==0.4.4
debugpy==1.4.1
decorator==5.0.9
entrypoints==0.3
Flask==2.0.1
gunicorn==20.1.0
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
nest-asyncio==1.5.1
nltk==3.6.2
numpy==1.18.5
packaging==20.4
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
# pywin32==301
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.8.28
six==1.16.0
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
typing-extensions==3.10.0.2
wcwidth==0.2.5
Werkzeug==2.0.1
